pyproject.toml
[tool.poetry]
name = "financetracker"
version = "0.1.0"
description = ""
authors = ["Nolan Gregory <65730528+nulzo@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-bootstrap5 = "^23.3"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
django-bootstrap-v5 = "^1.0.11"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
